// src/app/model/customer.model.ts
export interface Customer {
        "id": number;
        "firstName": String;
        "lastName": String;
        "age": number;
        "email": String;
        "contact": String;
        "createdBy": String;
        "createdDate": String;
        "updatedBy": String;
        "updatedDate": String
 }